+++ src/main/resources/locks_by_grant.sql
@@ -1,0 +1,4 @@
+SELECT relation::regclass as relname, granted, COUNT(*) as count
+FROM pg_locks
+WHERE relation::regclass::text ~ :relname
+group by relation, granted